backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.1

# AI/ML
openai-whisper==20231117
sentence-transformers==2.3.1
faiss-cpu==1.11.0

langchain==0.1.4
langchain-community==0.0.16

# Document processing
pymupdf==1.23.8
python-docx==1.1.0

# Audio/Video processing
ffmpeg-python==0.2.0
pydub==0.25.1

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
mongomock-motor==0.0.21
httpx==0.26.0
